# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API системы управления заявками на полеты с JWT аутентификацией",
        "title": "Flight Request API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/api/rumbs": {
            "get": {
                "description": "Возвращает JSON массив всех румбов в системе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rumbs"
                ],
                "summary": "Получить список всех румбов",
                "responses": {
                    "200": {
                        "description": "Список румбов",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ds.Rumb"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "description": "Создает новый румб в системе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rumbs"
                ],
                "summary": "Создать новый румб",
                "parameters": [
                    {
                        "description": "Данные для создания румба",
                        "name": "rumb",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.Rumb"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Созданный румб",
                        "schema": {
                            "$ref": "#/definitions/ds.Rumb"
                        }
                    },
                    "400": {
                        "description": "Неверные данные запроса",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/rumbs/{id}": {
            "get": {
                "description": "Возвращает детальную информацию о конкретном румбе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rumbs"
                ],
                "summary": "Получить информацию о румбе по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID румба",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о румбе",
                        "schema": {
                            "$ref": "#/definitions/ds.Rumb"
                        }
                    },
                    "400": {
                        "description": "Неверный ID румба",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Румб не найден",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "description": "Обновляет информацию о существующем румбе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rumbs"
                ],
                "summary": "Обновить информацию о румбе",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID румба",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные румба",
                        "name": "rumb",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.Rumb"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновленный румб",
                        "schema": {
                            "$ref": "#/definitions/ds.Rumb"
                        }
                    },
                    "400": {
                        "description": "Неверные данные запроса",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет румб из системы по ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rumbs"
                ],
                "summary": "Удалить румб",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID румба",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сообщение об успешном удалении",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "Проверяет учетные данные и возвращает JWT токен",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Аутентификация пользователя",
                "parameters": [
                    {
                        "description": "Учетные данные",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "JWT токен и информация о пользователе",
                        "schema": {
                            "$ref": "#/definitions/ds.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Неверные данные запроса",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Неверные учетные данные",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "/auth/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает информацию о текущем аутентифицированном пользователе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Получить текущего пользователя",
                "responses": {
                    "200": {
                        "description": "Информация о пользователе",
                        "schema": {
                            "$ref": "#/definitions/ds.User"
                        }
                    },
                    "401": {
                        "description": "Пользователь не аутентифицирован",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Пользователь не найден",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обновляет логин и пароль текущего пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Обновить данные текущего пользователя",
                "parameters": [
                    {
                        "description": "Новые данные пользователя",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.AuthRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сообщение об успешном обновлении",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Неверные данные запроса",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Пользователь не аутентифицирован",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Создает нового пользователя в системе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Регистрация нового пользователя",
                "parameters": [
                    {
                        "description": "Данные пользователя",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.User"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Сообщение об успешной регистрации",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Неверные данные пользователя",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/auth/validate": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Проверяет валидность JWT токена",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Проверить токен",
                "responses": {
                    "200": {
                        "description": "Информация о токене",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/flyRequests": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает список заявок с фильтрацией. Пользователи видят только свои заявки, модераторы - все.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FlyRequests"
                ],
                "summary": "Получить список заявок на полет",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Фильтр по статусу",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Фильтр по дате формирования после",
                        "name": "formedAfter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Фильтр по дате формирования до",
                        "name": "formedBefore",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список заявок",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создает новую заявку на полет для текущего пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FlyRequests"
                ],
                "summary": "Создать новую заявку на полет",
                "responses": {
                    "201": {
                        "description": "Созданная заявка",
                        "schema": {
                            "$ref": "#/definitions/ds.FlyRequest"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/flyRequests/current": {
            "get": {
                "description": "Возвращает ID текущей черновой заявки и количество румбов в ней",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FlyRequests"
                ],
                "summary": "Получить текущую заявку пользователя",
                "responses": {
                    "200": {
                        "description": "Информация о текущей заявке",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/flyRequests/rumbs/{flyRequestID}/{rumbID}": {
            "put": {
                "description": "Обновляет расстояние и скорость ветра для румба в заявке",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FlyRequestRumbs"
                ],
                "summary": "Обновить параметры румба в заявке",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки на полет",
                        "name": "flyRequestID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID румба",
                        "name": "rumbID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновленные данные",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Неверные данные запроса",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет связь между румбом и заявкой на полет",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FlyRequestRumbs"
                ],
                "summary": "Удалить румб из заявки",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки на полет",
                        "name": "flyRequestID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID румба",
                        "name": "rumbID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сообщение об успешном удалении",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Неверные ID",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/flyRequests/{id}": {
            "get": {
                "description": "Возвращает полную информацию о заявке на полет включая связанные румбы",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FlyRequests"
                ],
                "summary": "Получить детальную информацию о заявке на полет",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки на полет",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Детальная информация о заявке",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Неверный ID заявки",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Заявка не найдена",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "description": "Обновляет информацию о заявке на полет",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FlyRequests"
                ],
                "summary": "Обновить заявку на полет",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки на полет",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления заявки",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.FlyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновленная заявка",
                        "schema": {
                            "$ref": "#/definitions/ds.FlyRequest"
                        }
                    },
                    "400": {
                        "description": "Неверные данные запроса",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет заявку на полет по ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FlyRequests"
                ],
                "summary": "Удалить заявку на полет",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки на полет",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Перенаправление на главную страницу"
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "/fly_calculation/{id}": {
            "get": {
                "description": "Отображает HTML страницу с детальной информацией о заявке на полет",
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "Frontend"
                ],
                "summary": "Получить страницу заявки на полет",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки на полет",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "HTML страница заявки",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Неверный ID заявки",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Заявка не найдена",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ds.LoginRequest": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "ds.Rumb": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "Уникальный идентификатор направления",
                    "type": "integer",
                    "example": 1
                },
                "image": {
                    "description": "Ссылка на изображение направления",
                    "type": "string",
                    "example": "/images/north-east.png"
                },
                "title": {
                    "description": "Название направления",
                    "type": "string",
                    "example": "Север-Восток"
                }
            }
        },
...
        "handler.AuthRequest": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}